Correct the working directory for CGI scripts
According to http://www.ietf.org/rfc/rfc3875 the working directory for a CGI script invocation "SHOULD be set to the directory containing the script". Previously Jetty's CGI servlet instead used the CGI servlet's resourceBase, which does not account for the possibility of scripts existing in sub-directories of the resourceBase.

Fixes: #326
Signed-off-by: Matthew Sheppard <msheppard@funnelback.com>